% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputFilesFromDoGsOnly.R
\name{annotatetranscript}
\alias{annotatetranscript}
\title{Title}
\usage{
annotatetranscript(res)
}
\arguments{
\item{res}{}
}
\description{
Title
}
\examples{

y <- ThreeUTR:::annotatetranscript(res$reBR)

}
