% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{CheckBatch}
\alias{CheckBatch}
\title{CheckBatch}
\usage{
CheckBatch(dir.name, output_bacth_check, df.NT)
}
\arguments{
\item{df.NT}{}
}
\description{
CheckBatch
}
\examples{


df.NT<-re.DESeq.DoGs.plus.minus.gene.interested[,c(1,9:16)]
rownames(df.NT)<-df.NT[,1]
df.NT.2<- df.NT[,-1]

CheckBatch(dir.name,'Btach_using_DoGs',df.NT.2)

}
