% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{convert.z.score}
\alias{convert.z.score}
\title{Title}
\usage{
convert.z.score(z, two.sided = T)
}
\arguments{
\item{one.sided}{}
}
\description{
Title
}
\examples{
convert.z.score(1.454648e-04,two.sided=T)
convert.z.score(1.454648e-04,two.sided=F)

convert.z.score(-1.428500e-04,two.sided=T)
convert.z.score(-1.428500e-04,two.sided=F)
lapply(re.3UTR$re.FC.sorted$pvalue,function(x){
 y<-convert.z.score(x,two.side=F)
 y
 }
 )

one.sided.p<-convert.z.score(re.3UTR$re.FC.sorted$pvalue,two.sided=F)


}
