% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{convertbam2bed}
\alias{convertbam2bed}
\title{convertbam2bed}
\usage{
convertbam2bed(input.bamfile.dir, output.bedfile.dir)
}
\arguments{
\item{output.bedfile.dir}{}
}
\description{
convertbam2bed
}
\examples{

input.bamfile.dir <- '/media/H_driver/2016/Ramin_azhang/for_bioinfo_core/RNA_seq'
output.bedfile.dir <- '/Volumes/Bioinformatics$/Aimin_project'

res <- convertbam2bed(input.bamfile.dir,output.bedfile.dir)


R -e 'library(ChipSeq);library(ThreeUTR);re <- ThreeUTR:::convertbam2bed('/scratch/projects/bbc/aiminy_project/DoGs/BAM','/scratch/projects/bbc/aiminy_project/DoGs')'

}
