% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{convertBam2StrandBw2}
\alias{convertBam2StrandBw2}
\title{R -e 'library(ChipSeq);library(ThreeUTR);ThreeUTR:::convertBam2StrandBw2('/scratch/projects/bbc/aiminy_project/DoGs/BAM','/scratch/projects/bbc/aiminy_project/DoGs/BW_Perl_bg',BigMem=TRUE,cores=16)'}
\usage{
convertBam2StrandBw2(input.bam.file.dir, output.bw.file.dir, BigMem = FALSE,
  cores = 8, Memory = 36864, Wall.time = "72:00", span.ptile = 8)
}
\description{
R -e 'library(ChipSeq);library(ThreeUTR);ThreeUTR:::convertBam2StrandBw2('/scratch/projects/bbc/aiminy_project/DoGs/BAM','/scratch/projects/bbc/aiminy_project/DoGs/BW_Perl_bg',BigMem=TRUE,cores=16)'
}
