% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{convertCountFile2Table}
\alias{convertCountFile2Table}
\title{res <- convertCountFile2Table("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt")}
\usage{
convertCountFile2Table(input.count.file.dir, input.file.pattern)
}
\description{
res <- convertCountFile2Table("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt")
}
