% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{DEAnalysisAdjustByBatch}
\alias{DEAnalysisAdjustByBatch}
\title{DEAnalysisAdjustByBatch}
\usage{
DEAnalysisAdjustByBatch(df.NT, Nbatch)
}
\arguments{
\item{df.NT}{}
}
\description{
DEAnalysisAdjustByBatch
}
\examples{

df.NT<-re.DESeq.DoGs.plus.minus.gene.interested[,c(1,9:16)]

df.NT<-Re.unadjusted.adjusted$DE[,c(1,9:16)]

rownames(df.NT)<-df.NT[,1]
df.NT<- df.NT[,-1]

re.DoGs.adjusted.by.batch<-DEAnalysisAdjustByBatch(df.NT.2)

}
