% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{FilterTranscripts}
\alias{FilterTranscripts}
\title{FilterTranscripts}
\usage{
FilterTranscripts(input.dir, input.file.pattern)
}
\description{
FilterTranscripts
}
\examples{

Re.txs<-FilterTranscripts('/media/H_driver/2016/Ramin_azhang/for_bioinfo_core/RNA_seq/','final_list.csv')



}
