% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{gaoidtransform}
\alias{gaoidtransform}
\title{Title}
\usage{
gaoidtransform(input.file, output.file)
}
\arguments{
\item{output.file}{}
}
\description{
Title
}
\examples{
input.file <- '/Volumes/Bioinformatics$/Zhen-Gao/UTR/Gencode-3UTR-transcript.tsv'
output.file <- '/Volumes/Bioinformatics$/Zhen-Gao/UTR/Gencode-3UTR-transcript-transformed.tsv'
ThreeUTR:::gaoidtransform(input.file,output.file)

}
