% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getcounts}
\alias{getcounts}
\title{getcounts}
\usage{
getcounts(input.bamfile.dir, annotation.bed.file, ld, rd, output.count.file.dir)
}
\arguments{
\item{output.count.file.dir}{}
}
\description{
getcounts
}
\examples{

getcounts()



}
