% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{getcountsfromMatchedbed}
\alias{getcountsfromMatchedbed}
\title{getcountsfromMatchedbed}
\usage{
getcountsfromMatchedbed(input.bedfile.dir, output.count.file.dir, filter.sample)
}
\arguments{
\item{ld}{upstream base}

\item{rd}{downstream base}

\item{output.matched.bed.file.dir}{}
}
\description{
getcountsfromMatchedbed
}
\examples{

res <- matchbed2annotation(input.bedfile.dir,annotation.bed.file,
ld,rd,output.matched.bed.file.dir)

input.bedfile.dir <- res$output.bedfile.dir

res <- getcountsfromMatchedbed (input.bedfile.dir,output.count.file.dir)

}
