% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputFilesFromDoGsOnly.R
\name{getutrcount}
\alias{getutrcount}
\title{getutrcount}
\usage{
getutrcount(dir.name, input.file.pattern)
}
\arguments{
\item{input.file.pattern}{}
}
\description{
getutrcount
}
\examples{

dir.name <- "/Volumes/Bioinformatics$/Aimin_project/UTR/Counts/"
input.file.pattern <- "count.txt"

res <- ThreeUTR:::getutrcount(dir.name, input.file.pattern)

}
