% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{matchAndDE}
\alias{matchAndDE}
\title{res <- convertCountFile2Table("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt")
res.new <- ThreeUTR:::matchAndDE(res,file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),group.comparision = c("condition","Untreated","Treated"))}
\usage{
matchAndDE(res, sample.infor.file, group.comparision = c("condition",
  "Untreated", "Treated"))
}
\description{
res.new <- ThreeUTR:::matchAndDE(res,file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),group.comparision = c("condition","Treated","Untreated"))
}
