% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{matchbed2annotation}
\alias{matchbed2annotation}
\title{matchbed2annotation}
\usage{
matchbed2annotation(input.bedfile.dir, annotation.bed.file, ld, rd,
  output.matched.bed.file.dir)
}
\arguments{
\item{ld}{upstream base}

\item{rd}{downstream base}

\item{output.matched.bed.file.dir}{}
}
\description{
matchbed2annotation
}
\examples{

res <- convertbam2bed(input.bamfile.dir,output.bedfile.dir)
input.bedfile.dir <- res$output.bedfile.dir
annotation.bed.file <- annotation.bed.file
ld <- ld
rd <- rd
output.matched.bed.file.dir <- output.matched.bed.file.dir

res <- matchbed2annotation(input.bedfile.dir,annotation.bed.file,
ld,rd,output.matched.bed.file.dir)

}
