% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{MatchTxs2Gene.R}
\alias{MatchTxs2Gene.R}
\title{MatchTxs2Gene}
\usage{
MatchTxs2Gene.R(re.rMAT, Re.txs, dir.name, out.file)
}
\description{
MatchTxs2Gene
}
\examples{



dir.name='/media/H_driver/2016/Ramin_azhang/for_bioinfo_core/RNA_seq/Results4NewData/'
output.file='InterestedGeneTxsSorted.csv'


Re.interested.gene<-MatchTxs2Gene.R(re.rMAT,Re.txs,dir.name,output.file)


Re.interested.gene[which(Re.interested.gene$SYMBOL \%in\% c('XYLB')),]



}
