% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{norm_factors}
\alias{norm_factors}
\title{Title}
\usage{
norm_factors(mat)
}
\arguments{
\item{mat}{}
}
\description{
Title
}
\examples{


raw.count<-re.rMAT.com.total$DE[,10:17]

raw.count.dds<-counts(re.rMAT.com.total.test.2$dds)

raw.count.DE<-re.rMAT.com.total.test.2$DE[,c(1,10:17)]

raw.count.DE.2<-raw.count.DE[match(rownames(raw.count.dds),raw.count.DE$tx_name),]

head(raw.count.DE.2)

raw.count.2<-apply(raw.count,2,as.numeric)

head(raw.count)


nor.fac<-norm_factors(raw.count.2)

estimateSizeFactorsForMatrix(raw.count.2)


}
