% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{parserAnnotationFile}
\alias{parserAnnotationFile}
\title{parserAnnotationFile}
\usage{
parserAnnotationFile(input.annotation.file)
}
\description{
parserAnnotationFile
}
\examples{

input.annotation.file <- '/Volumes/Bioinformatics$/Aimin_project/UTR/3UTR_GFF2.txt'

y <- parserAnnotationFile(input.annotation.file)


}
