% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{parserreadfiles}
\alias{parserreadfiles}
\title{parserreadfiles}
\usage{
parserreadfiles(input.file.dir, input.file.type, sample.group = NULL,
  filter.sample = NULL)
}
\arguments{
\item{output.file.dir}{}
}
\description{
parserreadfiles
}
\examples{
input.file.dir <- '/media/H_driver/2016/Ramin_azhang/for_bioinfo_core/RNA_seq'
output.file.dir <- '/Volumes/Bioinformatics$/Aimin_project'


res <- parserreadfiles(input.file.dir,'bam',filter.sample='emp1')

}
