% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputFilesFrom5Cases.R
\name{ProcessOutputFilesFrom5Cases}
\alias{ProcessOutputFilesFrom5Cases}
\title{ProcessOutputFilesFrom5Cases}
\usage{
ProcessOutputFilesFrom5Cases(dir.name, input.file.pattern, out.dir.name,
  out.file.pattern, adjust_by_batch)
}
\arguments{
\item{dir.name:}{the path for input files}

\item{input.file.pattern:}{input file pattern}
}
\description{
read the mapping results from the Dogs of each gene
}
\examples{

dir.name="/media/aiminyan/DATA/Ramin_azhang/Counts5CasesEachSample/"

input.file.pattern="*count.2.txt"

out.dir.name="/media/aiminyan/DATA/Ramin_azhang/Counts5CasesEachSample/"
out.file.pattern= "DoGs_adjust_by_batch"


Re.unadjusted.adjusted<-ProcessOutputFilesFrom5Cases(dir.name,input.file.pattern,out.dir.name,out.file.pattern,adjust_by_batch="YES")

}

