% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{ProcessOutputFilesFromDoGsOnly}
\alias{ProcessOutputFilesFromDoGsOnly}
\title{ProcessOutputFilesFromDoGsOnly}
\usage{
ProcessOutputFilesFromDoGsOnly(dir.name, input.file.pattern, out.dir.name,
  out.file.pattern.interested, out.file.pattern.positive.gene,
  out.file.pattern.negative.gene, out.file.pattern.all, dir.name.gene.list,
  pattern.4.gene.list, adjust_by_batch, permutation.set.up = NULL)
}
\arguments{
\item{adjust_by_batch}{}
}
\description{
ProcessOutputFilesFromDoGsOnly
}
\examples{

dir.name='/media/aiminyan/DATA/Ramin_azhang/Counts4DoGsOnlyRmOne/'

input.file.pattern='*count.2.txt'

dir.name.gene.list='/media/H_driver/2016/Ramin_azhang/'
pattern.4.gene.list='final_list.csv'


out.dir.name='/media/aiminyan/DATA/Ramin_azhang/Counts4DoGsOnlyRmOne/'

for (i in 1:10) {

time.string = gsub(':', '-', gsub(' ', '_', as.character(Sys.time())))
out.dir.name=paste0('/media/aiminyan/DATA/Ramin_azhang/Counts4DoGsOnlyRmOnePermutationAt_',time.string,'/')
dir.create(out.dir.name)

out.file.pattern.interested='DoGs_adjust_by_batch_interested_gene'
out.file.pattern.positive.gene='DoGs_adjust_by_batch_positive'
out.file.pattern.negative.gene='DoGs_adjust_by_batch_negative'
out.file.pattern.all= 'DoGs_adjust_by_batch_all'

Re.unadjusted.adjusted<-ProcessOutputFilesFromDoGsOnly(dir.name,input.file.pattern,out.dir.name,out.file.pattern.interested,
out.file.pattern.positive.gene,
out.file.pattern.negative.gene,
out.file.pattern.all,
dir.name.gene.list,
pattern.4.gene.list,
adjust_by_batch='YES')

}

save.image(file=paste0(out.dir.name,'re_save_2.RData'))
savehistory(file=paste0(out.dir.name,'re_save_2.Rhistory'))
}
