% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{ReadGeneList}
\alias{ReadGeneList}
\title{Title}
\usage{
ReadGeneList(dir.name, input.file.pattern.4.interested.gene)
}
\arguments{
\item{input.file.pattern.4.interested.gene}{}
}
\description{
Title
}
\examples{
 dir.name='/media/H_driver/2016/Ramin_azhang/'
 input.file.pattern.4.interested.gene='final_list.csv'
gene.interested<-ReadGeneList(dir.name, input.file.pattern.4.interested.gene)

}
