% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDoGs.R
\name{runDoGs}
\alias{runDoGs}
\title{R -e 'library(ChipSeq);library(ThreeUTR);ThreeUTR:::runDoGs("SRP058633",file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),"/projects/ctsi/bbc/Genome_Ref/Homo_sapiens/UCSC/hg19/Annotation/Genes/genes.gtf","/projects/ctsi/bbc/Genome_Ref/Homo_sapiens/UCSC/hg19/Sequence/Bowtie2Index/genome","/projects/ctsi/bbc/aimin/annotation/","/scratch/projects/bbc/aiminy_project/DoGs/TestPipeline")'}
\usage{
runDoGs(sra.accession.number, sample.info.file, gene.gtf, genome.index,
  processed.gene.gtf, output.dir)
}
\description{
sample.info.file=file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt")
gene.gtf="/projects/ctsi/bbc/Genome_Ref/Homo_sapiens/UCSC/hg19/Annotation/Genes/genes.gtf"
genome.index="/projects/ctsi/bbc/Genome_Ref/Homo_sapiens/UCSC/hg19/Sequence/Bowtie2Index/genome"
processed.gene.gtf="/projects/ctsi/bbc/aimin/annotation/"
output.dir="/scratch/projects/bbc/aiminy_project/DoGs/TestPipeline"
}
