% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{splitBam}
\alias{splitBam}
\title{R -e 'library(ChipSeq);library(ThreeUTR);ThreeUTR:::splitBam('/scratch/projects/bbc/aiminy_project/DoGs/BAM','/scratch/projects/bbc/aiminy_project/DoGs/Bam_split')'}
\usage{
splitBam(input.bam.file.dir, output.bw.file.dir)
}
\description{
R -e 'library(ChipSeq);library(ThreeUTR);ThreeUTR:::splitBam('/scratch/projects/bbc/aiminy_project/DoGs/BAM','/scratch/projects/bbc/aiminy_project/DoGs/Bam_split')'
}
