% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{sumCount4Downstream}
\alias{sumCount4Downstream}
\title{library(org.Hs.eg.db)
res12 <- ThreeUTR:::sumCount4Downstream("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt",file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),c(1,2))
res21 <- ThreeUTR:::sumCount4Downstream("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt",file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),c(2,1))}
\usage{
sumCount4Downstream(input.count.file.dir, input.file.pattern, sample.infor.file,
  group.comparision = c("condition", 1, 2))
}
\description{
library(org.Hs.eg.db)
res12 <- ThreeUTR:::sumCount4Downstream("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt",file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),c(1,2))
res21 <- ThreeUTR:::sumCount4Downstream("~/Dropbox (BBSR)/Aimin_project/Research/DoGs/Counts","*.txt",file.path(system.file("extdata",package = "ThreeUTR"),"sample_infor.txt"),c(2,1))
}
