% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{TestUseLongestTxsOfGene}
\alias{TestUseLongestTxsOfGene}
\title{TestUseLongestTxsOfGene}
\usage{
TestUseLongestTxsOfGene(Re.unadjusted.adjusted, out.dir.name, out.file.pattern)
}
\arguments{
\item{Re.unadjusted.adjusted}{}
}
\description{
TestUseLongestTxsOfGene
}
\examples{


out.dir.name='/media/aiminyan/DATA/Ramin_azhang/Counts5CasesEachSample/'
out.file.pattern='DoGs_using_longest_txs_adjust_by_batch'

re.longest.transcript<-TestUseLongestTxsOfGene(Re.unadjusted.adjusted,out.dir.name,out.file.pattern)

}
