% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{useFastqDumpConvertSra2Fastq}
\alias{useFastqDumpConvertSra2Fastq}
\title{Example}
\usage{
useFastqDumpConvertSra2Fastq(sra.file.dir, output.dir, wait.job.name = NULL)
}
\description{
R -e 'library(ChipSeq);library(ThreeUTR);re <- ThreeUTR:::useFastqDumpConvertSra2Fastq("/nethome/axy148/DoGsExample","/scratch/projects/bbc/aiminy_project/DoGsFastq")'
}
\details{
On pegasus, you need to run this if you want to start conversion automatically after downloading using wget

R -e 'library(ChipSeq);library(ThreeUTR);re <- ThreeUTR:::useFastqDumpConvertSra2Fastq("/nethome/axy148/DoGsExample","/scratch/projects/bbc/aiminy_project/DoGsFastq",wait.job.name = "wgetDownload")'
}
