% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{useWget2Download}
\alias{useWget2Download}
\title{Example:
R -e 'library(ChipSeq);library(ThreeUTR);re <- ThreeUTR:::useWget2Download("SRP058633","/nethome/axy148/DoGsExample")'}
\usage{
useWget2Download(sra.accession.number, output.dir)
}
\description{
re <- ThreeUTR:::useWget2Download("SRP058633","/nethome/axy148/DoGsExample")
}
