% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{getcounts}
\alias{getcounts}
\title{getcounts}
\usage{
getcounts(input.bamfile.dir, annotation.bed.file, ld, rd, output.count.file.dir,
  filter.sample)
}
\arguments{
\item{output.count.file.dir}{}
}
\description{
getcounts
}
\examples{

getcounts()

R -e 'library(ChipSeq);library(ThreeUTR);re <- ThreeUTR:::getcounts('/scratch/projects/bbc/aiminy_project/DoGs/BAM','/projects/ctsi/bbc/aimin/annotation/hg19_DoGs_2.bed',0,0,'/scratch/projects/bbc/aiminy_project/DoGs')'

}
