% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR.R
\name{getutrcount}
\alias{getutrcount}
\title{getutrcount}
\usage{
getutrcount(dir.name, input.file.pattern, sample.infor.file,
  group.comparision = c(1, 2))
}
\arguments{
\item{input.file.pattern}{}
}
\description{
getutrcount
}
\examples{

dir.name <- '/Volumes/Bioinformatics$/Aimin_project/UTR/NewCounts'
input.file.pattern <- 'count.txt'

res <- ThreeUTR:::getutrcount(dir.name, input.file.pattern,file.path(system.file('extdata',package = 'ThreeUTR'),'sample_infor.txt'))

}
