% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputFilesFrom3UTR.R
\name{ProcessOutputFilesFrom3UTR}
\alias{ProcessOutputFilesFrom3UTR}
\title{ProcessOutputFilesFrom3UTR}
\usage{
ProcessOutputFilesFrom3UTR(dir.name, input.file.pattern)
}
\arguments{
\item{input_file_pattern}{}
}
\description{
ProcessOutputFilesFrom3UTR
}
\examples{

dir.name="/media/H_driver/2016/Ramin_azhang/for_bioinfo_core/RNA_seq/Results/"
input.file.pattern="*downstream.count.txt"

re.rMAT<-ProcessOutputFilesFrom3UTR(dir.name,input.file.pattern)



}

